move lstm to new structure, render pytorch/lstm optional, allow makefile posargs
