+++ .github/workflows/main.yml
@@ -243,10 +243,19 @@ jobs:
         os: [ 'ubuntu-latest' ]
         python-version: [ "3.10", "3.11", "3.12" ]
         julia: [ false ]
+        pytorch: [ false ]
+        pytest-args: [ "--ignore='docs/notebooks/extreme_value_analysis.ipynb' --ignore='docs/notebooks/lstm.ipynb'" ]
         include:
           - os: 'ubuntu-latest'
             python-version: "3.10"
             julia: true
+            pytorch: false
+            pytest-args: "--ignore='docs/notebooks/lstm.ipynb'"
+          - os: 'ubuntu-latest'
+            python-version: "3.10"
+            julia: false
+            pytorch: true
+            pytest-args: "--ignore='docs/notebooks/extreme_value_analysis.ipynb'"
     defaults:
       run:
         shell: bash -l {0}
@@ -295,6 +304,10 @@ jobs:
         if: ${{ matrix.julia == true }}
         run: |
           micromamba install -c conda-forge "pyjuliacall>=0.9.20"
+      - name: Install pytorch
+        if: ${{ matrix.pytorch == true }}
+        run: |
+          micromamba install -c conda-forge "pytorch>=2.10"
       - name: Install xHydro
         run: |
           python -m pip install --no-deps .
@@ -303,14 +316,9 @@ jobs:
           micromamba list
           echo ESMF_VERSION=$(cat $ESMFMKFILE | grep "ESMF_VERSION_STRING=" | awk -F= '{print $2}' | tr -d "'")
           python -m pip check || true
-      - name: Test Notebooks (no Julia)
-        if: ${{ matrix.julia == false }}
+      - name: Test Notebooks
         run: |
-          make test-notebooks-lax-noextremes
-      - name: Test Notebooks (Julia)
-        if: ${{ matrix.julia == true }}
-        run: |
-          make test-notebooks-lax
+          make test-notebooks-lax ARGS="${{ matrix.pytest-args }}"
 
 #  test-ESMF-source:
 #    name: Tests using ESMF from sources (Python${{ matrix.python-version }})
