+++ environment-dev.yml
@@ -28,7 +28,6 @@ dependencies:
   - spotpy >=1.6.2
   - stackstac >=0.5.1
   - statsmodels >=0.14.0
-  - tensorflow>=2.10
   - tqdm
   - xarray >=2023.11.0
   - xarray-spatial >=0.3.7
@@ -39,6 +38,8 @@ dependencies:
   # Julia
   # - julia >=1.6.0  # Works locally, breaks remotely
   # - pyjuliacall >=0.9.20
+  # LSTM
+  # - tensorflow >=2.10
   # Dev tools and testing
   - pip >=25.0
   - black ==25.1.0
