+++ environment.yml
@@ -28,7 +28,6 @@ dependencies:
   - spotpy >=1.6.2
   - stackstac >=0.5.1
   - statsmodels >=0.14.0
-  - tensorflow>=2.10
   - tqdm
   - xarray >=2023.11.0
   - xarray-spatial >=0.3.7
@@ -38,3 +37,5 @@ dependencies:
   - xvec
   # Julia
   # - pyjuliacall >=0.9.20
+  # LSTM
+  # - tensorflow >=2.10
